Fix reading bytes sometimes outside the current valid input-buffer
abs_ninput + ninput - d_offset is incorrect, because d_offset is
a relative value of the position of the tag not absolute and thus bits_left
gets really big values which leads to if (bits_left >= bits_needed) block to
always execute. Fixing this fixed the issue where when decoding a lot of blocks
the decoding would stop working at some point.